+++ src/constants/settings.js
@@ -13,7 +13,8 @@ const settings = {
   api: "api",
   menuBar: "menuBar",
   playBackControl: "playBackControl",
-  muteAds: "muteAds",
+  muteArtists: "muteArtists",
+  mutedArtists: "mutedArtists",
   apiSettings: {
     root: "apiSettings",
     port: "apiSettings.port",